refactor: split massa-api-export lib to multiple files
