+++ massa-api-exports/Cargo.toml
@@ -5,6 +5,10 @@ authors = ["Massa Labs <info@massa.net>"]
 edition = "2021"
 
 [dependencies]
+displaydoc = "0.2"
+thiserror = "1.0"
+jsonrpsee-core = { version = "0.16.2" }
+jsonrpsee-types = { version = "0.16.2" }
 serde = { version = "1.0", features = ["derive"] }
 strum = { version = "0.24", features = ["derive"] }
 
@@ -14,8 +18,12 @@ massa_time = { path = "../massa-time" }
 massa_models = { path = "../massa-models" }
 massa_final_state = { path = "../massa-final-state" }
 
-[dev-dependencies]
-serial_test = "0.10"
+massa_consensus_exports = { path = "../massa-consensus-exports" }
+massa_hash = { path = "../massa-hash" }
+massa_network_exports = { path = "../massa-network-exports" }
+massa_protocol_exports = { path = "../massa-protocol-exports" }
+massa_execution_exports = { path = "../massa-execution-exports" }
+massa_wallet = { path = "../massa-wallet" }
 
 # for more information on what are the following features used for, see the cargo.toml at workspace level
 [features]
