+++ massa-api-exports/src/slot.rs
@@ -1,0 +1,14 @@
+// Copyright (c) 2022 MASSA LABS <info@massa.net>
+
+use massa_models::{amount::Amount, slot::Slot};
+
+use serde::{Deserialize, Serialize};
+
+/// slot / amount pair
+#[derive(Debug, Deserialize, Serialize)]
+pub struct SlotAmount {
+    /// slot
+    pub slot: Slot,
+    /// amount
+    pub amount: Amount,
+}
