+++ massa-api/Cargo.toml
@@ -6,7 +6,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-displaydoc = "0.2"
 jsonrpsee = { version = "0.16.2", features = ["server", "macros"] }
 async-trait = "0.1.58"
 serde = { version = "1.0", features = ["derive"] }
@@ -14,7 +13,6 @@ serde_json = "1.0.87"
 tower-http = { version = "0.3.4", features = ["full"] }
 tower = { version = "0.4.13", features = ["full"] }
 hyper = "0.14.20"
-thiserror = "1.0"
 tokio = { version = "1.23", features = ["full"] }
 tokio-stream = { version = "0.1", features = ["sync"] }
 tracing = "0.1"
@@ -22,7 +20,6 @@ itertools = "0.10"
 parking_lot = { version = "0.12", features = ["deadlock_detection"] }
 # custom modules
 massa_consensus_exports = { path = "../massa-consensus-exports" }
-massa_hash = { path = "../massa-hash" }
 massa_api_exports = { path = "../massa-api-exports" }
 massa_models = { path = "../massa-models" }
 massa_network_exports = { path = "../massa-network-exports" }
